backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart